Move pbs_current_user
